requirements.txt
amqp==5.3.1
annotated-types==0.7.0
anyio==4.8.0
asgiref==3.8.1
async-timeout==5.0.1
attrs==25.1.0
autopep8==2.3.2
billiard==4.2.1
cachetools==5.5.2
celery==5.4.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cron-descriptor==1.4.5
Django==5.1.6
django-celery-beat==2.7.0
django-cors-headers==4.7.0
django-extensions==3.2.3
django-redis==5.4.0
django-rest-passwordreset==1.5.0
django-silk==5.3.2
django-timezone-field==7.1
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
drf-spectacular==0.28.0
exceptiongroup==1.2.2
factory_boy==3.3.3
Faker==35.2.0
flower==2.0.1
google-auth==2.38.0
google-genai==1.3.0
gprof2dot==2024.6.6
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
humanize==4.11.0
idna==3.10
inflection==0.5.1
iniconfig==2.0.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kombu==5.4.2
langchain-core==0.3.40
langchain-ollama==0.2.3
langsmith==0.3.11
Markdown==3.7
numpy==2.2.3
ollama==0.4.7
orjson==3.10.15
outcome==1.3.0.post0
packaging==24.2
pdf2image==1.17.0
pillow==11.1.0
pluggy==1.5.0
prometheus_client==0.21.1
prompt_toolkit==3.0.50
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycodestyle==2.12.1
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
PyJWT==2.9.0
PySocks==1.7.1
pytesseract==0.3.13
pytest==8.3.4
pytest-django==4.9.0
pytest-factoryboy==2.7.0
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
PyYAML==6.0.2
redis==5.2.1
referencing==0.36.2
requests==2.32.3
requests-toolbelt==1.0.0
rpds-py==0.23.1
rsa==4.9
selenium==4.28.1
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
sqlparse==0.5.3
tenacity==9.0.0
tomli==2.2.1
tornado==6.4.2
trio==0.28.0
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2025.1
uritemplate==4.1.1
urllib3==2.3.0
vine==5.1.0
wcwidth==0.2.13
websocket-client==1.8.0
websockets==14.2
whitenoise==6.9.0
wsproto==1.2.0
zstandard==0.23.0
